requirements.txt
nltk~=3.5
Flask~=1.1.2
Sastrawi~=1.0.1
mysql-connector-python~=8.0.21
PyMySQL~=0.10.1